ML/requirements.txt
numpy
pandas
matplotlib
ultralytics
opencv-python-headless
easyocr
Flask
pymongo
scikit-learn
python-dotenv